image_preprocess/requirements.txt
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
# For HTTP API (server.py)
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
